// src/types/CanvasMetadataDto.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { LayerMetadataDto } from "./LayerMetadataDto";

export type CanvasMetadataDto = { canvas_width: number, canvas_height: number, active_layer_id: string, version: string, layers: Array<LayerMetadataDto>, };